requirements.txt
Flask==2.3.3
Pillow==10.0.0
Werkzeug==2.3.7
pillow-heif==0.15.0 